add /public/ to .gitignore, make path with rubizza numbers and fix constans
