+++ 3529/2/.gitignore
@@ -1,1 +1,3 @@
 token.yaml
+
+/public/*
