+++ 3529/2/Lib/Utiles/image_loader.rb
@@ -1,7 +1,12 @@
 module ImageLoader
+  API_FILE_ID = 'https://api.telegram.org/botTOKEN/getFile?file_id=file_id_real'.freeze
+  FILE_PATH = 'https://api.telegram.org/file/botTOKEN/filepath'.freeze
+
   def load_image
-    file_id =  payload['photo'].first['file_id']
-    response = HTTParty.get("https://api.telegram.org/bot#{TOKEN}/getFile?file_id=#{file_id}")
-    "https://api.telegram.org/file/bot#{TOKEN}/#{response['result']['file_path']}"
+    API_FILE_ID.gsub!('TOKEN', TOKEN)
+    API_FILE_ID.gsub!('file_id_real', payload['photo'].first['file_id'])
+    response = HTTParty.get(API_FILE_ID)
+    API_FILE_ID.gsub!('TOKEN', TOKEN)
+    API_FILE_ID.gsub!('filepath', response['result']['file_path'])
   end
 end
